Changed events to only send notification to other player.
Also added the new deploy script, along with a deploy_production
script.
The deploy script uses the key in .deploykey. You can use
that to deploy
to some test happenings you are using.
If you have a seperate plugin that you use in
serious Happenings and
don't want to break, you can add that key to
.deploykey_production, and
deploy to that plugin using the
deploy_production script. This script
will ask for confirmation if you want to
deploy to those plugins.